Added HELP command
This iterates over the commands that are recognized, and displays them as acceptable commands. NotImplemented commands are identified and displayed with a - to indicate that it knows what they are and that they are not implemented.

Added the Unidentified command type for commands that are not expected to be supported. They are not displayed in the HELP list, and attempts to use them will return the NotImplemented,Superfluous code.