Added SYST, and STAT commands
These commands are fairly rudimentary at this point. Not a lot of status is captured. And SYST always (though theoretically configurable) uses "UNIX" as the System Type. Since regardless of if this is on Unix or Windows, I don't care about \r\n, and will always use \n, this is probably a moot point anywhere, and I doubt I'll get around to porting this to some Mainframe other OS.