Added ABOR command
This is ultimately not very useful as transfers are Asynchronous and there is no point where multiple sockets are checked simultaneously, so if you are in a transfer, nothing will be acknowledged by the command channel until the transfer is finished. I have found that no client I have located seems to cause the ABOR command to be sent to test this anyway. I will probably re-achitect so that the main loop just listens for availability for sending data/incoming data on the data channel and delegates handling it at some point but the initial get this written design does not account for that, so ABOR just dutifully sends its 226 - Data connection closed (and closes the data connection) and that is all.