FTP-13 Added the EPSV Command
This is the "modern" mechanism for specifying the connect to address. This works in IPv4 and IPv6 mode (LPSV should theoretically support IPv4, but I cannot find a client that will use it if you use IPv4)

EPSV also required supporting the ALL and (net protocol) arguments. When ALL is used, any other data-opening-command becomes invalid. So they were all changed to return a Bad Command Sequence if they are called after a EPSV ALL.

Net protocols are handled by if there is a preferred specified then open a listen to all interfaces on port port. And obviously error if the specified network protocol is not supported, listing the supported protocols