FTP-6 Added AUTH command
Added the AUTH command and configured it so that it will work with TLS. Right now it only works with TLS/1.2+, and there are no other mechanisms which are supported and there are no plans to support any other mechanism. Without the PBSZ and PROT commands, this will not support the DATA channel.