FTP-7 Added CCC command
Added the CCC command. This allows transitioning the command channel back to unencrypted after the the AUTH command has been issued. The theoretical use for this is that Firewalls can sniff the FTP channel when it is unencrypted and allow expected ports to go through, while still allowing credentials to be sent over encrypted data. Most free FTP clients do not support the CCC command.

Part of performing this included finding out that that the DataChannelProtectionLevel stays in place. And that it can only be negotiated while the command channel is encrypted. So replacements had to be made to whether those commands are allowed. Also since the PROT command must only immediately follow the PBSZ command I added a check for that. As life has become difficult identifying what the last command issued was, I have added the last command to the Connection. This I accomplished by having commands identify themselves, which was done by implementing the Name() function to the FTPCommand interface. So All commands had to be modified to support this.