FTP-5 Added FEAT command
Added the FEAT command and new interface for extended commands. This will allow commands to identify if they are featured, and what kind of additional parameters they may allow (e.g. AUTH TLS)

Add commands that are beyond the RFC 959 set have been adapted to support this interface. Commands that do not support this interface are ignore by the FEAT command.

The security options ADAT, MIC, CONF, ENC are left to the security mechanism to determine if they are displayed (because they exist but are not used by TLS auth). And CCC only displays if CCC is enabled.