FTP-15 Added OPTS command
Added support for the OPTS command. None of the existing commands support options, so this doesn't actually do much. This will probably have to be revised as it looks like there are new commands that are official commands that will use options, and there are commands that offer no value, but do have options. If they are real commands they will have to be retrievable by the Command Map, but if they are not "real" commands, they will don't have to be, but they probably should.